# apps/heater-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Heater API (LT Capstone)
  description: API Server that controls the Heater via GPIO with PID Logic.
  version: 0.1.0
paths:
  /channels:
    get:
      tags:
      - Channels
      summary: List Channels
      description: 'List all configured channels from config.yaml.


        Returns information about all channels including their configuration

        and current runtime status (whether PID is active).'
      operationId: listChannels
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelListResponse'
  /channels/{channel_id}:
    get:
      tags:
      - Channels
      summary: Get Channel Info
      description: 'Get information about a specific channel.


        Returns configuration and runtime status for the specified channel.'
      operationId: getChannelInfo
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /channels/status/all:
    get:
      tags:
      - Channels
      summary: Get All Channels Status
      description: 'Get comprehensive PID status for all enabled channels at once.


        Returns detailed PID status including temperature, duty cycle, parameters,

        and error statistics for each enabled channel.'
      operationId: getAllChannelsStatus
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllChannelsStatus'
  /config/{channel_id}/target-temp:
    get:
      tags:
      - Config
      summary: Get Target Temp
      description: Get the current target temperature for a specific channel.
      operationId: getTargetTemp
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetTemp'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Config
      summary: Set Target Temp
      description: Set a new target temperature for a specific channel.
      operationId: setTargetTemp
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetTemp'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/{channel_id}/pid-parameters:
    get:
      tags:
      - Config
      summary: Get Pid Parameters
      description: Get the current PID parameters for a specific channel.
      operationId: getPidParameters
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parameters'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Config
      summary: Set Pid Parameters
      description: Update PID parameters for a specific channel.
      operationId: setPidParameters
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Parameters'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parameters'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/{channel_id}/all:
    get:
      tags:
      - Config
      summary: Get All Config
      description: Get all configuration (target temperature and PID parameters) for
        a specific channel.
      operationId: getAllConfig
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigAll'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pid/{channel_id}/start:
    post:
      tags:
      - PID
      summary: Start Pid
      description: Start the PID controller for a specific channel.
      operationId: startPID
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Startpid
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pid/{channel_id}/stop:
    post:
      tags:
      - PID
      summary: Stop Pid
      description: Stop the PID controller for a specific channel.
      operationId: stopPID
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Stoppid
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pid/{channel_id}/status:
    get:
      tags:
      - PID
      summary: Get Pid Status
      description: 'Get comprehensive PID status for a specific channel.


        Returns:

        - channel_id: Channel identifier

        - channel_name: Channel name

        - is_active: Whether PID controller is currently running

        - target: Target temperature

        - duty_cycle: Current PWM duty cycle

        - current_temp: Current temperature reading

        - pid_parameters: Current PID coefficients (Kp, Ki, Kd)

        - pid_variables: Internal PID variables (integral, last_error, last_measurement)

        - error_stats: Error statistics including counts for last 1m, 10m, and since
        start'
      operationId: getPidStatus
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PidStatusOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AllChannelsStatus:
      properties:
        channels:
          additionalProperties:
            $ref: '#/components/schemas/PidStatusOut'
          type: object
          title: Channels
      type: object
      required:
      - channels
      title: AllChannelsStatus
      description: Status of all channels at once.
    ChannelInfo:
      properties:
        channelId:
          type: string
          title: Channelid
        name:
          type: string
          title: Name
        gpioPin:
          type: integer
          title: Gpiopin
        sensorChannel:
          type: integer
          title: Sensorchannel
        enabled:
          type: boolean
          title: Enabled
        isActive:
          type: boolean
          title: Isactive
      type: object
      required:
      - channelId
      - name
      - gpioPin
      - sensorChannel
      - enabled
      - isActive
      title: ChannelInfo
      description: Channel information with runtime status.
    ChannelListResponse:
      properties:
        channels:
          items:
            $ref: '#/components/schemas/ChannelInfo'
          type: array
          title: Channels
      type: object
      required:
      - channels
      title: ChannelListResponse
      description: List of all configured channels.
    ConfigAll:
      properties:
        targetTemp:
          type: number
          title: Targettemp
        pidParameters:
          $ref: '#/components/schemas/Parameters'
      type: object
      required:
      - targetTemp
      - pidParameters
      title: ConfigAll
      description: 'Schema for all configuration data.


        Contains target temperature and PID parameters.'
    ErrorStats:
      properties:
        errorsLast1M:
          type: integer
          title: Errorslast1M
        errorsLast10M:
          type: integer
          title: Errorslast10M
        errorsSinceStart:
          type: integer
          title: Errorssincestart
        lastErrorMessage:
          anyOf:
          - type: string
          - type: 'null'
          title: Lasterrormessage
      type: object
      required:
      - errorsLast1M
      - errorsLast10M
      - errorsSinceStart
      - lastErrorMessage
      title: ErrorStats
      description: 'Schema for error statistics.


        Tracks API call errors over different time periods.'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Parameters:
      properties:
        kp:
          type: number
          title: Kp
          default: 1.0
        ki:
          type: number
          title: Ki
          default: 0.0
        kd:
          type: number
          title: Kd
          default: 0.0
      type: object
      title: Parameters
      description: 'Schema for temperature parameters.


        Used by GET and PUT /temp/parameters endpoints

        to retrieve or update temperature parameters.'
    PidStatusOut:
      properties:
        channelId:
          type: string
          title: Channelid
        channelName:
          type: string
          title: Channelname
        isActive:
          type: boolean
          title: Isactive
        target:
          type: number
          title: Target
        dutyCycle:
          type: number
          title: Dutycycle
        currentTemp:
          anyOf:
          - type: number
          - type: 'null'
          title: Currenttemp
        pidParameters:
          $ref: '#/components/schemas/Parameters'
        pidVariables:
          $ref: '#/components/schemas/PidVariables'
        errorStats:
          $ref: '#/components/schemas/ErrorStats'
      type: object
      required:
      - channelId
      - channelName
      - isActive
      - target
      - dutyCycle
      - currentTemp
      - pidParameters
      - pidVariables
      - errorStats
      title: PidStatusOut
      description: 'Schema for PID status output.


        Contains comprehensive information about PID controller state.'
    PidVariables:
      properties:
        integral:
          type: number
          title: Integral
        lastError:
          type: number
          title: Lasterror
        lastMeasurement:
          anyOf:
          - type: number
          - type: 'null'
          title: Lastmeasurement
      type: object
      required:
      - integral
      - lastError
      - lastMeasurement
      title: PidVariables
      description: 'Schema for PID internal variables.


        Contains the interval variables used in PID calculations.'
    TargetTemp:
      properties:
        target:
          type: number
          title: Target
      type: object
      required:
      - target
      title: TargetTemp
      description: 'Schema for target temperature.


        Used by GET and PUT /temp/target-temp endpoints

        to retrieve or update target temperature.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError